requirements.txt
asgiref==3.4.1
Brotli==1.0.9
cssselect2==0.4.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.9
django-appconf==1.0.5
django-crispy-forms==1.13.0
django-defender==0.9.1
djangorestframework==3.13.1
fonttools==4.27.1
html5lib==1.1
importlib-metadata==4.8.3
packaging==21.3
pkg-resources==0.0.0
pycparser==2.21
pydyf==0.1.2
pyparsing==3.0.7
pyphen==0.11.0
pytz==2021.3
redis==4.1.4
six==1.16.0
sqlparse==0.4.2
tinycss2==1.1.1
typing-extensions==4.0.0
webencodings==0.5.1
wrapt==1.13.3
zipp==3.6.0
zopfli==0.1.9
